Add assert helpers as separated file
Ref: https://github.com/hl7au/au-fhir-core-inferno/issues/71
